backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.6
weasyprint==63.1
python-multipart==0.0.20
pytest==9.0.2
pypdf==6.7.1
pydantic-settings==2.12.0
PyMuPDF==1.27.1
python-docx==1.2.0
httpx==0.28.1
